update libs (merge some function in Qmaps.py) + test the spectrum pipeline
